+++ web/modules/custom/ai_screening/src/Exception/InvalidArgumentException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Drupal\ai_screening\Exception;
+
+/**
+ * An invalid argument exception.
+ */
+class InvalidArgumentException extends Exception {
+
+}