+++ web/modules/custom/ai_screening/src/Helper/UserHelper.php
@@ -51,7 +51,7 @@ final class UserHelper extends AbstractHelper implements EventSubscriberInterfac
   /**
    * {@inheritdoc}
    */
-  public static function getSubscribedEvents() {
+  public static function getSubscribedEvents(): array {
     return [
       EntityHookEvents::ENTITY_PRE_SAVE => 'preSave',
     ];