+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+- [PR-18](https://github.com/itk-dev/ai-screening/pull/18)
+  Added submission data to track
 - [PR-20](https://github.com/itk-dev/ai-screening/pull/20)
   Add site branding block to navigation
   Remove field_ui from `core.extension.yml` and remove a field_ui related config file